Add files via upload
This is the first version of Return of the King Manual for Archipelago NOW PLAYABLE!!! 


*****NOTE*****
This currently does NOT have support for Co-op. There will probably be a separate AP world for co-op and that will be a project for a later date 

In order to get the full experience of the Randomizer you will need to complete Crack of Doom and/or Palantirs and then get every character to level 20 WITHOUT purchasing any upgrades whatsoever. Once you complete crack of doom then cheat codes are available and you can use them to level up the characters quickly!!! 



Setup Guide
1. Drag and drop the manual stable and Return of the King AP world Files into Archipelago/lib/worlds folder

2. Open the Archipelagolauncher.exe

3. click on generate Template Settings

4. In archipelago/players/templates ther will be the Return of the King YAML or you can grab this one from here as well 

5. configure the YAML however you want

6. Connect to Archipelago as you would through a normal game